Tweaks to support GHC 7.8 type checker.  1. Disable unused ParIMap class  2. add an ::Int ascription

+++ haskell/lvish/Data/Concurrent/SkipListMap.hs
@@ -353,7 +353,7 @@ debugShow (Slice (SLMap index lmbot) mstart mend) =
                          ] ]
     loop (Index indm slm) = do
       ls <- LM.toList indm
-      strs <- forM [ (i,tup) | i <- [0..] | tup@(k,_) <- ls ] $ -- , startCheck k
+      strs <- forM [ (i,tup) | i <- [(0::Int)..] | tup@(k,_) <- ls ] $ -- , startCheck k
               \ (ix, (key, (shortcut::t, val))) -> do
         -- Peek at the next layer down:
 {-
